+++ tutorials/hair_geometry/hair_geometry.cpp
@@ -5,7 +5,7 @@
 #include "../common/tutorial/benchmark_render.h"
 
 #if defined(EMBREE_SYCL_TUTORIAL)
-#  define NAME "hair_geometry_sycl (beta)"
+#  define NAME "hair_geometry_sycl"
 #  define FEATURES FEATURE_RTCORE | FEATURE_SYCL
 #else
 #define NAME "hair_geometry"