+++ tutorials/motion_blur_geometry/motion_blur_geometry.cpp
@@ -5,7 +5,7 @@
 #include "../common/tutorial/benchmark_render.h"
 
 #if defined(EMBREE_SYCL_TUTORIAL)
-#  define NAME "motion_blur_geometry_sycl (beta)"
+#  define NAME "motion_blur_geometry_sycl"
 #  define FEATURES FEATURE_RTCORE | FEATURE_SYCL
 #else
 #  define NAME "motion_blur_geometry"