+++ doc/src/changelog.md
@@ -2,6 +2,7 @@ Version History
 ---------------
 
 ### Embree 4.1.1
+-   SYCL version of Embree with GPU support is no longer in beta phase.
 -   Improved BVH build performance on many core machines for applications that oversubscribe threads.
 -   Added rtcGetGeometryTransformFromScene API function that can get used inside SYCL kernels.
 -   No longer linking to ze_loader in SYCL mode to avoid Intel(R) oneAPI Level Zero dependency